+++ api/src/main/java/net/jqwik/api/arbitraries/StringArbitrary.java
@@ -114,6 +114,6 @@ public interface StringArbitrary extends Arbitrary<String> {
 	 *
 	 * @param repeatProbability Must be between 0 (included) and 1 (excluded)
 	 */
-	@API(status = EXPERIMENTAL, since = "1.5.3")
+	@API(status = MAINTAINED, since = "1.5.3")
 	StringArbitrary repeatChars(double repeatProbability);
 }