+++ api/src/main/java/net/jqwik/api/lifecycle/PerProperty.java
@@ -22,7 +22,7 @@ import static org.apiguardian.api.API.Status.*;
 @Target({ElementType.ANNOTATION_TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @AddLifecycleHook(PerProperty.PerPropertyHook.class)
-@API(status = EXPERIMENTAL, since = "1.2.4")
+@API(status = MAINTAINED, since = "1.2.4")
 public @interface PerProperty {
 
 	interface Lifecycle {